+++ link/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     implementation project(':stripe-core')
     implementation project(':payments-ui-core')
     implementation project(':financial-connections')
+//    implementation project(':stripe-ui-core')
 
     implementation "androidx.appcompat:appcompat:$androidxAppcompatVersion"
     implementation "androidx.constraintlayout:constraintlayout:$androidxConstraintlayoutVersion"
