+++ link/src/main/java/com/stripe/android/link/ui/LinkTerms.kt
@@ -8,8 +8,8 @@ import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import com.stripe.android.link.R
-import com.stripe.android.ui.core.elements.Html
 import com.stripe.android.ui.core.paymentsColors
+import com.stripe.android.uicore.text.Html
 
 @Preview
 @Composable
