+++ payments-ui-core/build.gradle
@@ -7,6 +7,7 @@ apply plugin: 'org.jetbrains.kotlin.plugin.serialization'
 dependencies {
     implementation project(":stripe-core")
     implementation project(":payments-core")
+    implementation project(":stripe-ui-core")
     compileOnly project(":stripecardscan")
     compileOnly "com.google.android.libraries.places:places:$placesVersion"
