+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/AffirmElementUI.kt
@@ -9,6 +9,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import com.stripe.android.ui.core.R
 import com.stripe.android.ui.core.paymentsColors
+import com.stripe.android.uicore.text.Html
+import com.stripe.android.uicore.text.EmbeddableImage
 
 @Composable
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
