+++ payments-ui-core/src/main/java/com/stripe/android/ui/core/elements/AfterpayClearpayElementUI.kt
@@ -15,6 +15,8 @@ import com.stripe.android.ui.core.R
 import com.stripe.android.ui.core.elements.AfterpayClearpayHeaderElement.Companion.isClearpay
 import com.stripe.android.ui.core.paymentsColors
 import com.stripe.android.ui.core.shouldUseDarkDynamicColor
+import com.stripe.android.uicore.text.Html
+import com.stripe.android.uicore.text.EmbeddableImage
 
 @Composable
 @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
